MiniTeste

Description

Este módulo contém definições Haskell para o cálculo de funções recursivas simples.

Documentation

f :: [[Int]] -> [Int]

A função ’f’ calcula a soma dos elementos de uma lista e forma uma lista com esses elementos (f [[n,n1],[n2,n3]] retorna uma lista em que os seus elementos são as somas entre os valores das listas,ou seja, [n+n1,n2+n3]).

Definição dos tipos :

>>> :: Int
>>> ->Int

Exemplos de utilização:

>>> f [[]]
[]
>>> f [[1],[1,2,3],[3,-4,1]]
 [1,13]